gamelogic-1: seperated game logic into their respective files
