+++ app/src/app/pages/game/logic/Board.ts
@@ -1,0 +1,67 @@
+import { Application, Container, Rectangle, Sprite, Texture } from "pixi.js"
+
+export class Board extends Container {
+
+    private app: Application
+    private tileSize: number
+    private tileNumber: number
+    private tileScale: number
+    private tileTexture: Texture
+    private animationDuration: number
+    public boardSize: number
+
+    constructor(
+        app: Application,
+        tileSize: number = 256,
+        tileNumber: number = 8,
+        tileScale: number = 0.3,
+        tileTexture: Texture = Texture.from('assets/tile_water.png'),
+        animationDuration = 0.2
+    ) {
+        super()
+
+        this.app = app
+        this.tileSize = tileSize
+        this.tileNumber = tileNumber
+        this.tileScale = tileScale
+        this.boardSize = tileSize * tileScale * tileNumber
+        this.tileTexture = tileTexture
+        this.animationDuration = animationDuration
+
+        for (let i = 0; i < this.tileNumber; i++) {
+            for (let j = 0; j < this.tileNumber; j++) {
+                let child = Sprite.from(this.tileTexture)
+                child.anchor.set(0.5, 0.5)
+                child.scale.set(this.tileScale)
+                child.position.set(i * this.tileSize * tileScale, j * this.tileSize * tileScale)
+                child.eventMode = 'static'
+                child.addEventListener('click', e => {
+                    (e.currentTarget as Sprite).renderable = false
+                })
+                child.addEventListener('mouseover', e => {
+                    gsap.to((e.currentTarget as Sprite).scale, { x: this.tileScale + 0.05, y: this.tileScale + 0.05, duration: this.animationDuration })
+                })
+                child.addEventListener('mouseout', e => {
+                    gsap.to((e.currentTarget as Sprite).scale, { x: this.tileScale, y: this.tileScale, duration: this.animationDuration })
+                })
+                this.addChild(child)
+            }
+        }
+    }
+
+    public centerBoardVertically() {
+        this.position.y = (this.app.renderer.screen.bottom - this.tileSize * this.tileScale * (this.tileNumber-1)) / 2
+    }
+
+    public setLeftPadding(padding: number) {
+        this.position.x = this.tileSize / 2 * this.tileScale + padding
+    }
+
+    public getDimensions() {
+        return new Rectangle(
+            this.position.x - this.tileSize * this.tileScale / 2,
+            this.position.y - this.tileSize * this.tileScale / 2,
+            this.width + this.tileSize * this.tileScale,
+            this.height + this.tileSize * this.tileScale)
+    }
+}
